==优化目录结构 ==退出登录时, 自动清除业务数据(清空Model层) ==事件系统的Key改成int, 减少gc

+++ Client/Assets/Game/YouYouFramework/Managers/Audio/AudioManager.cs
@@ -43,10 +43,10 @@ namespace YouYou
             BGMSource.name = "BGMSource";
             BGMSource.outputAudioMixerGroup = GameEntry.Instance.MonsterMixer.FindMatchingGroups("BGM")[0];
 
-            GameEntry.PlayerPrefs.AddEventListener(PlayerPrefsDataMgr.EventName.MasterVolume, RefreshMasterVolume);
-            GameEntry.PlayerPrefs.AddEventListener(PlayerPrefsDataMgr.EventName.BGMVolume, RefreshBGM);
-            GameEntry.PlayerPrefs.AddEventListener(PlayerPrefsDataMgr.EventName.AudioVolume, RefreshAudio);
-            GameEntry.PlayerPrefs.AddEventListener(PlayerPrefsDataMgr.EventName.GamePause, OnGamePause);
+            GameEntry.PlayerPrefs.AddEventListener((int)PlayerPrefsDataMgr.EventName.MasterVolume, RefreshMasterVolume);
+            GameEntry.PlayerPrefs.AddEventListener((int)PlayerPrefsDataMgr.EventName.BGMVolume, RefreshBGM);
+            GameEntry.PlayerPrefs.AddEventListener((int)PlayerPrefsDataMgr.EventName.AudioVolume, RefreshAudio);
+            GameEntry.PlayerPrefs.AddEventListener((int)PlayerPrefsDataMgr.EventName.GamePause, OnGamePause);
 
             RefreshMasterVolume(null);
             RefreshBGM(null);
